requirement.txt
psycopg2
wechat_sdk
django-memcached
python-memcached